backend/foodgram/requirements.txt
django==4.2.6
djangorestframework==3.14.0
djoser==2.2.0
drf-yasg==1.21.7
pillow==10.0.1
django-filter==23.3